+++ server/routes/posts.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import { getPosts, createPost } from '../controllers/posts.js';
+import { getPosts, createPost } from '../controllers/postsController.js';
 
 const router = express.Router();
